dom/virtualfilesystem/test/mochitest/test_virtualfilesystem_request_timeout.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1187223
-->
<head>
<meta charset="utf-8">
<title>Test Inputport Connection Event</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="virtualfilesystem_common.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187223">Test FileSystemProvider Event
</a>
<script type="application/javascript;version=1.8">
"use strict";

SimpleTest.waitForExplicitFinish();

var Ci = SpecialPowers.Ci;

function testMount() {
  return new Promise(function(aResolve, aReject) {
    navigator.fileSystemProvider.mount(mountOption).then(
      function() {
        ok(true, "mount ok.");
        aResolve();
      },
      function(aError) {
        ok(false, "Fail to mount: " + aError);
        aResolve();
      }
    );
  });
}

function getVirtualFileSystem(aFileSystemId) {
  let vfsService = SpecialPowers.Cc["@mozilla.org/virtualfilesystem/virtualfilesystem-service;1"].getService(Ci.nsIVirtualFileSystemService);
  return vfsService.getVirtualFileSystemById(aFileSystemId);
}

function testRequestTimeout() {
  return new Promise(function(aResolve, aReject) {
    var fileSystem = getVirtualFileSystem(mountOption.fileSystemId);
    ok(true, "must have the fileSystem");
    let callback = {
      onSuccess: function(requestId, value, hasMore) {
      },
      onError: function(requestId, error) {
        is(error, Ci.nsIVirtualFileSystemCallback.ERROR_TIME_OUT, "Got expected error code.");
        aResolve();
      }
    };
    fileSystem.abort(0, callback);
  });
}

function testUnmount() {
  return new Promise(function(aResolve, aReject) {
    navigator.fileSystemProvider.unmount(unmountOption).then(
      function() {
        ok(true, "unmount ok.");
        aResolve();
      },
      function(aError) {
        ok(false, "Fail to unmount: " + aError);
        aResolve();
      }
    );
  });
}

function runTest() {
  ok(navigator.fileSystemProvider, "should have navigator.fileSystemProvider");
  testMount()
  .then(testRequestTimeout)
  .then(testUnmount)
  .then(function() {
    info('test finished');
    SimpleTest.finish();
  });
}

SpecialPowers.pushPrefEnv({"set": [["dom.filesystemprovider.enabled", true],
                                   ["dom.filesystemprovider.request_timeout", 5000],
                                    //to ignore app scope check.
                                   ["dom.ignore_webidl_scope_checks", true],
                                   ["dom.mozBrowserFramesEnabled", true],
                                   ["dom.ipc.tabs.disabled", false]]}, function() {
  SpecialPowers.pushPermissions(
    [{type: "browser", allow: true, context: document },
     {type: 'filesystemprovider', allow: true, context: document},], runTest);
});
</script>
</body>
</html>
